move example to doc folder
