+++ Code/ToshibaAir/src/main.h
@@ -25,7 +25,7 @@
 #include <ArduinoJson.h>
 #include <NTPClient.h> 
 #include <WiFiUdp.h>
-#include "toshiba_serial.hpp"
+#include "..\doc\examples\air_test\toshiba_serial.hpp"
 #include "MySimpleTimer.hpp"
 #include "process_request.hpp"
